#18 refactor: sepertate TrendingProducts component from Home component
